[FIX] #217 seperated component
